app/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.7
boto3==1.28.57
python-dotenv==1.0.0
SQLAlchemy==1.4.53
Flask-SQLAlchemy==2.5.1